New class Habitat
This class *will* handle the habitat of an individual
By habitat we mean something like "comprehensive environment",
a habitat will consist of at least
  * chemical molecules (possibly proteins but could also be proteins
    or anything else one would come up with)
  * the current phenotypic target (which we will no longer call
    environment because this term is ambiguous)
  * an entity that will manage the evolution of the habitat over time
    (e.g. variation and noise applied to the phenotypic target but
    also the set of possible phenotypic targets and the rules that
    define how and when we switch from one to another -- this last
    example is mostly, or maybeeven *only* for use in RAevol)

+++ src/libaevol/ae_stat_record.h
@@ -219,9 +219,9 @@ class ae_stat_record
     int32_t  _nb_bases_non_essential_including_nf_genes;
     
     #ifdef __REGUL
-      double  _nb_influences;
-      double  _nb_enhancing_influences;
-      double  _nb_operating_influences;
+      int32_t  _nb_influences;
+      int32_t  _nb_enhancing_influences;
+      int32_t  _nb_operating_influences;
       double  _av_value_influences;
       double  _av_value_enhancing_influences;
       double  _av_value_operating_influences;
